stoa-operator/requirements-dev.txt
-r requirements.txt
ruff>=0.4
black>=24.0
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.12
pytest-cov>=5.0
pytest-httpx>=0.30
